:recycle: :fire: refactoring

+++ src/controllers/user.controller.ts
@@ -1,17 +1,24 @@
-import { NextFunction, Request, Response } from 'express';
-import { BadQueryStringException } from '../exceptions/badQueryString.exception';
-import { RecordNotFoundException } from '../exceptions/recordNotFound.exception';
-import { RecordNotUpdatedException } from '../exceptions/recordNotUpdated.exception';
-import { getPagination, getPagingData } from '../helpers/pagination.helper';
-import { UserCreationAttributes, userCreationScheme } from '../models/user.model';
 import UsersRepository from '../repositories/users.repository';
+import { BadRequestException } from '../exceptions/bad-request.exception';
+import { IUserCreationAttributes, UserCreationSchema } from '../interfaces/user.interface';
+import { NextFunction, Request, Response } from 'express';
+import { RecordNotFoundException } from '../exceptions/record-not-found.exception';
+import { RecordNotUpdatedException } from '../exceptions/record-not-updated.exception';
+import { StatusCodes } from 'http-status-codes';
+import { getPagination, getPagingData } from '../helpers/pagination.helper';
+import { Joi, requestValidator } from '../helpers/request-validator.helper';
 
 export class UserController {
   static async create(req: Request, res: Response, next: NextFunction) {
     try {
-      const payload: UserCreationAttributes = await userCreationScheme.validateAsync(req.body);
-      const data = await UsersRepository.create(payload);
-      return res.json({ data });
+      const { body: payload } = await requestValidator(
+        {
+          body: UserCreationSchema,
+        },
+        req,
+      );
+      const row = await UsersRepository.create(payload);
+      return res.status(StatusCodes.CREATED).json({ data: row });
     } catch (error) {
       next(error);
     }
@@ -19,8 +26,8 @@ export class UserController {
 
   static async readAll(req: Request, res: Response, next: NextFunction) {
     try {
-      const data = await UsersRepository.readAll();
-      return res.json({ data });
+      const rows = await UsersRepository.readAll();
+      return res.json({ data: rows });
     } catch (error) {
       next(error);
     }
@@ -28,16 +35,26 @@ export class UserController {
 
   static async readAllWithPagination(req: Request, res: Response, next: NextFunction) {
     try {
-      const { name = '', page = '0', size = '10' } = req.query;
-      if (typeof name !== 'string') throw new BadQueryStringException('name');
-      if (typeof page !== 'string') throw new BadQueryStringException('page');
-      if (typeof size !== 'string') throw new BadQueryStringException('size');
+      const querySchema = Joi.object({
+        name: Joi.string(),
+        page: Joi.string(),
+        size: Joi.string(),
+      });
+
+      const {
+        query: { name = '', page = '0', size = '10' },
+      } = await requestValidator(
+        {
+          query: querySchema,
+        },
+        req,
+      );
 
       const { limit, offset } = getPagination(page, size);
-      const filter: UserCreationAttributes = { name };
-      const data = await UsersRepository.readAllWithPaginationAndFilter(limit, offset, filter);
-      const response = getPagingData(data, page, limit);
-      return res.json(response);
+      const filter: IUserCreationAttributes = { name };
+      const { rows, count } = await UsersRepository.readAllWithPaginationAndFilter(limit, offset, filter);
+      const { totalItems, data, totalPages, currentPage } = getPagingData(rows, count, page, limit);
+      return res.json({ totalItems, data, totalPages, currentPage });
     } catch (error) {
       next(error);
     }
@@ -45,12 +62,21 @@ export class UserController {
 
   static async readById(req: Request, res: Response, next: NextFunction) {
     try {
-      const { id } = req.params;
-      const data = await UsersRepository.readById(id);
-      if (data === null) {
-        throw new RecordNotFoundException(id);
-      }
-      return res.json({ data });
+      const paramsSchema = Joi.object({
+        id: Joi.string().guid().required(),
+      });
+      const {
+        params: { id },
+      } = await requestValidator(
+        {
+          params: paramsSchema,
+        },
+        req,
+      );
+
+      const row = await UsersRepository.readById(id);
+      if (row === null) throw new RecordNotFoundException(id);
+      return res.json({ data: row });
     } catch (error) {
       next(error);
     }
@@ -58,12 +84,24 @@ export class UserController {
 
   static async update(req: Request, res: Response, next: NextFunction) {
     try {
-      const payload: UserCreationAttributes = await userCreationScheme.validateAsync(req.body);
-      const { id } = req.params;
+      const paramsSchema = Joi.object({
+        id: Joi.string().guid().required(),
+      });
+      const {
+        body: payload,
+        params: { id },
+      } = await requestValidator(
+        {
+          body: UserCreationSchema,
+          params: paramsSchema,
+        },
+        req,
+      );
+
+      const row = await UsersRepository.readById(id);
+      if (row === null) throw new RecordNotFoundException(id);
       const data = await UsersRepository.update(id, payload);
-      if (data === null) {
-        throw new RecordNotUpdatedException(id);
-      }
+      if (data === null) throw new RecordNotUpdatedException(id);
       return res.json({ data });
     } catch (error) {
       next(error);
@@ -72,9 +110,22 @@ export class UserController {
 
   static async delete(req: Request, res: Response, next: NextFunction) {
     try {
-      const { id } = req.params;
-      const data = await UsersRepository.delete(id);
-      return res.json({ data });
+      const paramsSchema = Joi.object({
+        id: Joi.string().guid().required(),
+      });
+      const {
+        params: { id },
+      } = await requestValidator(
+        {
+          params: paramsSchema,
+        },
+        req,
+      );
+
+      const row = await UsersRepository.readById(id);
+      if (row === null) throw new BadRequestException();
+      await UsersRepository.delete(id);
+      return res.status(StatusCodes.NO_CONTENT);
     } catch (error) {
       next(error);
     }
